Update torch_dtype in DebugCLIPTextModel instantiation from float32 to float16 for optimized performance. Mark subproject commit in pipeline/compel as dirty to indicate uncommitted changes.

+++ pipeline/loader.py
@@ -34,7 +34,7 @@ def load_pipeline_components():
     )
 
     # Original Transformers Models
-    original_clip_l = models.DebugCLIPTextModel.from_pretrained(defaults.DEFAULT_BASE_MODEL, subfolder="text_encoder", torch_dtype=torch.float32).to(device)
+    original_clip_l = models.DebugCLIPTextModel.from_pretrained(defaults.DEFAULT_BASE_MODEL, subfolder="text_encoder", torch_dtype=torch.float16).to(device)
     original_clip_g = models.DebugCLIPTextModelWithProjection.from_pretrained(defaults.DEFAULT_BASE_MODEL, subfolder="text_encoder_2", torch_dtype=torch.float16).to(device)
     
     compel_original = Compel(
